% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kestrelbox-package.R
\docType{package}
\name{kestrelbox-package}
\alias{kestrelbox}
\alias{kestrelbox-package}
\title{kestrelbox: Data Package with Kestrel Nesting Box Data}
\description{
Loads two datasets containing kestrel nesting box data for Pennsylvania and New Jersey. One dataset (bybox) is organized so that each row represents a nesting box, and the other (bychick) is organized so that each row represents a chick.
}
\author{
\strong{Maintainer}: Leo Soccio \email{leo.soccio@gmail.com}

}
\keyword{internal}
